requirements.txt
python-telegram-bot==20.3
openai==1.14.2
faiss-cpu
flask
